pyproject.toml
[project]
name = "khata"
version = "0.0.1"
description = "Open-source trading journal for Indian brokers. Self-hosted, broker-synced, attachment-friendly."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Swathi" }]
keywords = ["trading", "journal", "dhan", "zerodha", "india", "nse", "options"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.7",
    "apscheduler>=3.10",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "respx>=0.21",
]

[project.scripts]
khata = "khata.cli:app"

[project.entry-points."khata.adapters"]
dhan = "khata.adapters.dhan:DhanAdapter"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["khata"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
